Lives, Push Ability, Better building breakage, Pickups
Added three Lives as well as a pickup entity that works for life
pickups but soon for various other ones as well.

I added the ability to apply an outward force on all objects nearby the
ship as a way to help the player when surrounded by objects. This will
be able to be a pick up with a one time burst or pickups and maybe a
fuel bar that depletes with use.

The buildings that break into pieces when hit not have an initial
velocity that matches the players from impact plus a bit of randomness
as well as the initial angular velocity

Added pickups and have a life pickup that will soon be used for other
types of pickups as well